Create js file for handlehabits and add to app.js
