+++ src/App.js
@@ -5,10 +5,11 @@ import List from './components/List/List'
 import Calendar from './components/Calendar/Calendar'
 import Navigation from './components/Navigation/Navigation'
 import Header from './components/Header/Header'
+import handleHabits from './handleHabits'
 import { v4 as uuid } from 'uuid'
 
 function App() {
-  const [habits, setHabits] = useState([])
+  const { addHabit, toggleCompleted, removeHabit } = handleHabits()
 
   return (
     <Router>
@@ -17,17 +18,13 @@ function App() {
         <Switch>
           <Route exact path="/">
             <List
-              habits={habits}
               onItemClick={toggleCompleted}
               headline={'Dashboard'}
+              onDelete={removeHabit}
             />
           </Route>
           <Route path="/create">
-            <Form
-              habits={habits}
-              onSubmit={addHabit}
-              headline={'Create a habit'}
-            />
+            <Form onSubmit={addHabit} headline={'Create a habit'} />
           </Route>
           <Route path="/overview">
             <Calendar headline={'Your monthly overview'} />
@@ -37,18 +34,6 @@ function App() {
       <Navigation />
     </Router>
   )
-  function addHabit(name) {
-    setHabits([...habits, { name, completed: false, id: uuid() }])
-  }
-
-  function toggleCompleted(id) {
-    const index = habits.findIndex((habit) => habit.id === id)
-    const habit = habits[index]
-    const before = habits.slice(0, index)
-    const after = habits.slice(index + 1)
-    const updatedHabit = { ...habit, completed: !habit.completed }
-    setHabits([...before, updatedHabit, ...after])
-  }
 }
 
 export default App
