+++ src/util.js
@@ -18,9 +18,3 @@ export function displayToday() {
     today.getDate() + ' ' + months[today.getMonth()] + ' ' + today.getFullYear()
   return date
 }
-
-// function removeHabit(id) {
-//   const index = habits.findIndex((habit) => habit.id === id)
-//   setHabits([...habits.slice(0, index), ...habits.slice(index + 1)])
-//   return removeHabit
-// }
